Refactor image builders
